fix?(main): only set_x outputs when a state change occurs
This fixes prob, since prob is only replaced (and there the seed is
reset to default) when prob is changed.

Bug: The sequence is always the same for each Prob::PXX value.
Fix: Set the seed "more appropriately", whatever that ends up being.

Bug: When rate changes, that output breaks. This has to do with count
getting off cycle.
Fix: Maybe do the internal state setting better in `seq`, or queue up
changes in `seq` and only apply then on a major tick. In any case, not
the binary's concern.