fix(euclid): caret animates up to x96
The fix is to only send `StateChange::Index` when we're on the euclid
screen, and only for that single screen. Two other `StateChange`
variants need index information because they impact whether or not the
euclid screen is going to get shown; OutputType change and NextScreen.

There are still bugs: the caret will not update until the next index
change when switching to euclid. We'll squash that bug in a future
commit.