chore(pointer): redraw screen
There is still a bug in that next element does not have index. Two
possible fixes: also pass through index in tick from seq for
`NextElement` state change messages, or keep track of prev_element in
state and pass that through with the state change.

Going to see how the latter goes.